use checkPermission instead of hasAdminManagementPermissions

+++ cibseven-webclient-core/src/main/java/org/cibseven/webapp/rest/AdminService.java
@@ -73,7 +73,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	@Parameter(description = "Specifies the field to sort by") @RequestParam Optional<String> sortBy,
 	@Parameter(description = "Specifies the order of the sorting") @RequestParam Optional<String> sortOrder,
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.USER, PermissionConstants.READ_ALL);
+		checkPermission(user, SevenResourceType.USER, PermissionConstants.READ_ALL);
 		return bpmProvider.findUsers(id, firstName, firstNameLike, lastName, lastNameLike, 
 				email, emailLike, memberOfGroup, memberOfTenant, idIn, firstResult, maxResults, sortBy, sortOrder, user);
 	}
@@ -94,7 +94,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void createUser(
 			@RequestBody NewUser user,
 			Locale loc, CIBUser flowUser) {
-		hasAdminManagementPermissions(flowUser, SevenResourceType.USER, PermissionConstants.CREATE_ALL);
+		checkPermission(flowUser, SevenResourceType.USER, PermissionConstants.CREATE_ALL);
 		bpmProvider.createUser(user, flowUser);
 	}
 
@@ -117,7 +117,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			@RequestBody User user, 
 			Locale loc, CIBUser flowUser) {
-		hasAdminManagementPermissions(flowUser, SevenResourceType.USER, PermissionConstants.UPDATE_ALL);
+		checkPermission(flowUser, SevenResourceType.USER, PermissionConstants.UPDATE_ALL);
 		bpmProvider.updateUserProfile(userId, user, flowUser);
 	}
 	
@@ -140,7 +140,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Group Id") @PathVariable String groupId, 
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			Locale loc, CIBUser flowUser) {
-		hasAdminManagementPermissions(flowUser, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
+		checkPermission(flowUser, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
 		bpmProvider.addMemberToGroup(groupId, userId, flowUser);
 	}
 	
@@ -163,7 +163,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Group Id") @PathVariable String groupId, 
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			Locale loc, CIBUser flowUser) {
-		hasAdminManagementPermissions(flowUser, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
+		checkPermission(flowUser, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteMemberFromGroup(groupId, userId, flowUser);
 	}
 	
@@ -188,7 +188,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			@RequestBody Map<String, Object> data, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.USER, PermissionConstants.UPDATE_ALL);
+		checkPermission(user, SevenResourceType.USER, PermissionConstants.UPDATE_ALL);
 		bpmProvider.updateUserCredentials(userId, data, user);
 	}
 
@@ -207,7 +207,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void deleteUser(
 			@Parameter(description = "User Id") @PathVariable String userId, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.USER, PermissionConstants.DELETE_ALL);
+		checkPermission(user, SevenResourceType.USER, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteUser(userId, user);
 	}
 
@@ -233,7 +233,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Specifies the maximum number of results to return"
 					+ "<br>Will return less results if there are no more results left") @RequestParam Optional<String> maxResults,		
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.GROUP, List.of("ALL", "READ"));
+		checkPermission(user, SevenResourceType.GROUP, PermissionConstants.READ_ALL);
 		String decodedMember = member.map(value -> {
 			try {
 				return URLDecoder.decode(value, StandardCharsets.UTF_8.toString());
@@ -282,7 +282,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void createGroup(
 			@RequestBody UserGroup group,
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.GROUP, PermissionConstants.CREATE_ALL);
+		checkPermission(user, SevenResourceType.GROUP, PermissionConstants.CREATE_ALL);
 		bpmProvider.createGroup(group, user);
 	}
 
@@ -305,7 +305,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Group Id") @PathVariable String groupId, 
 			@RequestBody UserGroup group, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.GROUP, PermissionConstants.UPDATE_ALL);
+		checkPermission(user, SevenResourceType.GROUP, PermissionConstants.UPDATE_ALL);
 		bpmProvider.updateGroup(groupId, group, user);
 	}
 
@@ -324,7 +324,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void deleteGroup(
 			@Parameter(description = "Group Id") @PathVariable String groupId, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
+		checkPermission(user, SevenResourceType.GROUP, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteGroup(groupId, user);
 	}
 	
@@ -352,7 +352,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Maximum number of results to return"
 					+ "<br>Will return less results if there are no more results left") @RequestParam Optional<String> maxResults,
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.AUTHORIZATION, PermissionConstants.READ_ALL);
+		checkPermission(user, SevenResourceType.AUTHORIZATION, PermissionConstants.READ_ALL);
 		return bpmProvider.findAuthorization(id, type, userIdIn, groupIdIn, resourceType, resourceId,
 				sortBy, sortOrder, firstResult, maxResults, user);
 	}
@@ -380,7 +380,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public ResponseEntity<Authorization> createAuthorization(
 			@RequestBody Authorization authorization,
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.AUTHORIZATION, PermissionConstants.CREATE_ALL);
+		checkPermission(user, SevenResourceType.AUTHORIZATION, PermissionConstants.CREATE_ALL);
 		return bpmProvider.createAuthorization(authorization, user);
 	}
 
@@ -409,7 +409,7 @@ public class AdminService extends BaseService implements InitializingBean {
 			@Parameter(description = "Authorization Id") @PathVariable String authorizationId, 
 			@RequestBody Map<String, Object> data, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.AUTHORIZATION, PermissionConstants.UPDATE_ALL);
+		checkPermission(user, SevenResourceType.AUTHORIZATION, PermissionConstants.UPDATE_ALL);
 		bpmProvider.updateAuthorization(authorizationId, data, user);
 	}
 
@@ -428,7 +428,7 @@ public class AdminService extends BaseService implements InitializingBean {
 	public void deleteAuthorization(
 			@Parameter(description = "Authorization Id") @PathVariable String authorizationId, 
 			Locale loc, CIBUser user) {
-		hasAdminManagementPermissions(user, SevenResourceType.AUTHORIZATION, PermissionConstants.DELETE_ALL);
+		checkPermission(user, SevenResourceType.AUTHORIZATION, PermissionConstants.DELETE_ALL);
 		bpmProvider.deleteAuthorization(authorizationId, user);
 	}
